requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.65.2
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.3.2
